(TK-47)(enh) - Refacto serializers and models
