+++ gardeniq/hardware/protocols/errors.py
@@ -3,11 +3,13 @@ from gardeniq.base.utils import GardenEnum
 
 class FrameParsingError(Exception):
     """Raised when frame parsing fails."""
+
     pass
 
 
 class FrameProcessingError(Exception):
     """Raised when frame processing fails."""
+
     pass
 
 
@@ -26,6 +28,7 @@ class CommandError(GardenEnum):
         CHECKSUM_ERR: Data integrity check failed, indicating potential data corruption
         DEV_NOT_READY: Device is not in a state to accept or execute commands
     """
+
     UNKNOW_CMD = "UNKNOW_CMD"
     INVALID_PARAM = "INVALID_PARAM"
     TIMEOUT = "TIMEOUT"
