+++ gardeniq/orderlg/views.py
@@ -5,6 +5,7 @@ from gardeniq.orderlg.models import Order
 from gardeniq.orderlg.serializers import ArgumentReadOnlySerializer
 from gardeniq.orderlg.serializers import ArgumentSerializer
 from gardeniq.orderlg.serializers import OrderDetailReadOnlySerializer
+from gardeniq.orderlg.serializers import OrderListReadOnlySerializer
 from gardeniq.orderlg.serializers import OrderSerializer
 
 
@@ -16,6 +17,7 @@ class ArgumentAPIModelView(DisableAPIViewMixin, BaseAPIModelViewSet):
 
 class OrderAPIModelView(DisableAPIViewMixin, BaseAPIModelViewSet):
     serializer_class = OrderSerializer
+    list_serializer_class = OrderListReadOnlySerializer
     detail_serializer_class = OrderDetailReadOnlySerializer
     queryset = Order.objects.all()
